Update TOC.yml
Simplified the TOC tree structure with the logic of 5 questions, QDK, Q#, Techniques in QDK (including libraries), quantum computing theoretical concepts, etc. This inevitably moves the libraries down in order. However, the way to increase visibility of libraries is not literally moving it up in order. To increase their visibility, we need to direct usage of libraries from examples. Let's dive into the Q# examples next and identify the libraries that are being used. If there's not enough usage, we need to create more examples using these libraries.