Describe second qubit in 'prepare entanglement'
For some reason only the first qubit being set to the initial state is mentioned - to make things less confusing, it is worth adding that the second qubit is set to Zero. 

This helps to understand the next sentences (why the second qubit will always be the same as what we measure for the first).